# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/invite:
    get:
      operationId: inviteMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileInviteResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Generate a mobile invite link
      tags:
      - mobile
  /api/v1/invite/rotate:
    post:
      operationId: rotateMobileInvite
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileInviteResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Explicitly rotate the mobile invite link password
      tags:
      - mobile
  /api/v1/mobile/disable:
    post:
      operationId: disableMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Disable the Connect Mobile LAN bridge
      tags:
      - mobile
  /api/v1/mobile/enable:
    post:
      operationId: enableMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Enable the Connect Mobile LAN bridge and issue a fresh password
      tags:
      - mobile
  /api/v1/mobile/regenerate:
    post:
      operationId: regenerateMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Rotate the Connect Mobile password, dropping any connected phone
      tags:
      - mobile
  /api/v1/mobile/status:
    get:
      operationId: getMobileStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
      summary: Check whether Connect Mobile's LAN bridge is enabled
      tags:
      - mobile
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    MobileInviteResponse:
      properties:
        link:
          type: string
        password:
          type: string
      required:
      - link
      - password
      type: object
    MobileStatusResponse:
      properties:
        enabled:
          type: boolean
        host:
          type: string
        password:
          type: string
        port:
          type: integer
        warning:
          type: string
      required:
      - enabled
      - host
      - port
      - password
      - warning
      type: object
...